requirements.txt
appdirs==1.4.3
astroid==1.5.2
colorama==0.3.9
cycler==0.10.0
flake8==3.3.0
flake8-polyfill==1.0.1
h5py==2.6.0
isort==4.2.5
lazy-object-proxy==1.2.2
mando==0.6.4
matplotlib==2.0.0
mccabe==0.6.1
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.21.0
pip==7.0.3
Pillow==4.0.0
protobuf==3.2.0
pycodestyle==2.3.1
pyflakes==1.5.0
pylint==1.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
radon==2.0.2
scikit-learn==0.19.0
scipy==0.18.1
six==1.10.0
tensorflow-gpu==1.0.1
tqdm==4.11.2
wheel==0.29.0
wrapt==1.10.10
